Move methods out of OrchestratorSettings and leave as props only
